astrology-engine/requirements.txt
# ==========================================
# ASTROAI - PYTHON ASTROLOGY ENGINE
# Python Dependencies (requirements.txt)
# ==========================================
# Installation: pip install -r requirements.txt
# Python Version: 3.11+
# ==========================================

# ==========================================
# WEB FRAMEWORK
# ==========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==========================================
# ASTROLOGY CALCULATIONS
# ==========================================
# Swiss Ephemeris - NASA-grade astronomical calculations
pyswisseph==2.10.3.2

# Timezone handling
pytz==2023.3
python-dateutil==2.8.2

# ==========================================
# GEOCODING & LOCATION
# ==========================================
# For converting place names to lat/long coordinates
geopy==2.4.1
timezonefinder==6.2.0

# ==========================================
# HTTP REQUESTS
# ==========================================
httpx==0.25.2
requests==2.31.0

# ==========================================
# DATA PROCESSING
# ==========================================
numpy==1.26.2
pandas==2.1.4

# ==========================================
# DATABASE (if using Python for storage)
# ==========================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# ==========================================
# CACHING
# ==========================================
redis==5.0.1

# ==========================================
# LOGGING
# ==========================================
python-json-logger==2.0.7

# ==========================================
# ENVIRONMENT VARIABLES
# ==========================================
python-dotenv==1.0.0

# ==========================================
# SECURITY
# ==========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# ==========================================
# VALIDATION & PARSING
# ==========================================
python-multipart==0.0.6

# ==========================================
# CORS & MIDDLEWARE
# ==========================================
# Already included in fastapi

# ==========================================
# TESTING (Optional - for development)
# ==========================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# ==========================================
# CODE QUALITY (Optional - for development)
# ==========================================
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# ==========================================
# PERFORMANCE MONITORING (Optional)
# ==========================================
# prometheus-client==0.19.0

# ==========================================
# NOTES
# ==========================================
# 1. Install all: pip install -r requirements.txt
# 2. For production, freeze versions: pip freeze > requirements-freeze.txt
# 3. Create virtual environment first:
#    python -m venv venv
#    source venv/bin/activate  (Linux/Mac)
#    venv\Scripts\activate     (Windows)
# 4. Swiss Ephemeris data files needed:
#    Download from: https://www.astro.com/ftp/swisseph/ephe/
#    Required files: sepl_*.se1, semo_*.se1, seas_*.se1
# 5. Update regularly: pip list --outdated
# 6. Security check: pip-audit (install separately)

# ==========================================
# SWISS EPHEMERIS DATA DOWNLOAD
# ==========================================
# After installing pyswisseph, download ephemeris data:
# 
# Method 1 - Manual Download:
# wget https://www.astro.com/ftp/swisseph/ephe/sepl_18.se1
# wget https://www.astro.com/ftp/swisseph/ephe/semo_18.se1
# wget https://www.astro.com/ftp/swisseph/ephe/seas_18.se1
#
# Method 2 - Python script:
# import swisseph as swe
# swe.set_ephe_path('./ephemeris_data')
# 
# Place files in: ./ephemeris_data/ directory
# Update EPHEMERIS_PATH in .env file

# ==========================================
# INSTALLATION COMMANDS
# ==========================================
# 
# Basic installation:
# pip install -r requirements.txt
#
# Development installation (includes testing tools):
# pip install -r requirements.txt
#
# Production installation (without dev tools):
# pip install fastapi uvicorn pyswisseph pytz geopy sqlalchemy redis
#
# Upgrade all packages:
# pip install --upgrade -r requirements.txt
#
# Check installed versions:
# pip list
#
# Security audit:
# pip install pip-audit
# pip-audit

# ==========================================
# TROUBLESHOOTING
# ==========================================
# 
# If pyswisseph installation fails:
# - Windows: Install Visual C++ Build Tools
# - Linux: sudo apt-get install build-essential
# - Mac: xcode-select --install
#
# If psycopg2 fails:
# - Try: pip install psycopg2-binary instead
#
# If numpy/pandas fails:
# - Upgrade pip: pip install --upgrade pip
# - Install wheel: pip install wheel